third_party/WebKit/LayoutTests/http/tests/inspector/sources/debugger/source-map-http-header-expected.txt

Tests that SourceMap and X-SourceMap http headers are propagated to scripts in the front-end.

Reloading...
Added script:
  url: http://127.0.0.1:8000/inspector/sources/debugger/resources/source-map.php
  sourceMapUrl: http://localhost/source.map.js
Added script:
  url: http://127.0.0.1:8000/inspector/sources/debugger/resources/x-source-map.php
  sourceMapUrl: http://localhost/x-source.map.js
Page reloaded.
Finished reload.
